Extract container with fake stamp service to a trait
